修复从干簧管切换到震动开关时，休眠无法解除的问题 修复屏保计时器的一些隐藏问题 修复分钟转秒的时候发生精度损失的问题

+++ MathFun.cpp
@@ -135,7 +135,7 @@ double ESP32_ADC2Vol(uint16_t adc) {
  * @param {uint32_t} min
  * @return {*}
  */
-uint32_t Minute2Millis(uint32_t min) {
+uint32_t Minute2Millis(float min) {
     return min * 60 * 1000;
 }
 /***
